backend/game-move/requirements.txt
psycopg2-binary==2.9.9
chess==1.10.0